requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
h5py>=3.8.0
gdown>=4.7.1
pandas>=1.5.0
